convert format

+++ src/file_util.h
@@ -2,9 +2,9 @@
 #include "nums.h"
 
 typedef struct {
-    i32 w; // TODO make unsigned
-    i32 h;
-    u32 pitch;
+    u16 w; // TODO make unsigned
+    u16 h;
+    u16 pitch;
     i32 comp;
     u8 *image; // TODO maybe make this the MmAllocateContiguousMemoryEx part only and delete this array after fully loaded textureaddress?
     // u16 w; // TODO make unsigned
